agents/requirements.txt
# Python dependencies for uAgents
uagents>=0.21.0
uagents-core>=0.2.0

# OpenAI API for AI processing
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Async support
asyncio
aiohttp

# Existing dependencies (keep these)
requests>=2.31.0
